Create Audit Tables for CircleCI job
The CircleCI Job build_data_dictionary, creates a database.
It does not create the Audit tables because SPRING_JPA_HIBERNATE_DDL_AUTO
is not set to create, create-drop or update

This means running migrations on an audit table e.g. hearing_outcome_aud,
will result in an error because it was does not exist.