Migrate to Swagger 3
Use OpenAPIConfiguration settings to set the API Documentation details
and provide a security requirement of Bearer token - specifically
HMPPS Auth tokens.

This means that the API requests require authentication with HMPPS Auth

Also, the API Documentation / the service is behind an IP Allowlist
which is an added layer of security.