PIC-3671 Reintroduce data-dictionary  (#1198)
* Readd data-dictionary

This allows the analytics team and members of the probation in court
team to view the database schema in a user friendly UI.

* Update docker push of data_dictionary image with latest tag

We aren't pushing up the latest tag. The data_dictionary Helm deployment
is using the latest tag for the image. This means it's using an older image
to deploy on.