Remove pagination from JPA / database layer
This is because we need to get data about all hearing outcomes
before paginating it. We need to get all the assigned users first then paginate.

This fixes a bug where the UI (prepare-a-case) users could not see a
list of all assigned users in order to filter on all records.