Paginate hearing outcomes in the code layer
First, get all assigned users on the filtered hearing defendant outcomes
results. Use that in the return as `assignedUsers` value in HearingOutcomeCaseList
Then paginate the results and fill in the rest of the HearingOutcomeCaseList
object data.

Create a V2 version of fetch hearing outcomes as the return object is different for
HearingOutcomeCaseList.

The V2 HearingOutcomeCaseList object will contain counts by hearing outcome
states in the filter section of the response. Hence, create getV2OutcomeCountsByState function