Add models used to structure the v2 response for /hearing-defendant-outcomes
Add models to describe how the /v2/courts/:code/hearing-defendant-outcomes
will return with 2 top level properties: records and filters.

The response will be a v2/HearingOutcomeCaseList.kt

The filter result will include all the possible ways hearing-defendant-outcomes
can be filtered. This will be used by the UI to dynamically create filters
. It'll mean the UI will automatically include a new filter option
if the back end enables it. This allows the UI to have less knowledge of
this API.

the original v1 HearingOutcomeCaseList.kt will have a new property
assignedUsers which contains ids of all assigned users on the hearing outcomes.