PIC-3725 and PIC-3808 Return all Assigned Users for court & Create V2 endpoint (#1194)
* Remove pagination from JPA / database layer

This is because we need to get data about all hearing outcomes
before paginating it. We need to get all the assigned users first then paginate.

This fixes a bug where the UI (prepare-a-case) users could not see a
list of all assigned users in order to filter on all records.

* Fix HearingOutcomeRepositoryCustom tests

Fix HearingOutcomeRepositoryCustom tests after changing the response
structure

* Paginate hearing outcomes in the code layer

First, get all assigned users on the filtered hearing defendant outcomes
results. Use that in the return as `assignedUsers` value in HearingOutcomeCaseList
Then paginate the results and fill in the rest of the HearingOutcomeCaseList
object data.

Create a V2 version of fetch hearing outcomes as the return object is different for
HearingOutcomeCaseList.

The V2 HearingOutcomeCaseList object will contain counts by hearing outcome
states in the filter section of the response. Hence, create getV2OutcomeCountsByState function

* Create PaginatedHeadersService

This will add custom headers to a response entity that includes
pagination information.

This will be used by the V2 endpoints that paginate data

Add test for PaginatedHeadersService

* Add models used to structure the v2 response for /hearing-defendant-outcomes

Add models to describe how the /v2/courts/:code/hearing-defendant-outcomes
will return with 2 top level properties: records and filters.

The response will be a v2/HearingOutcomeCaseList.kt

The filter result will include all the possible ways hearing-defendant-outcomes
can be filtered. This will be used by the UI to dynamically create filters
. It'll mean the UI will automatically include a new filter option
if the back end enables it. This allows the UI to have less knowledge of
this API.

the original v1 HearingOutcomeCaseList.kt will have a new property
assignedUsers which contains ids of all assigned users on the hearing outcomes.

* Add FilterHearingDefendantOutcomesService

Add functionality that adds all the available filters for hearing
defendant outcomes e.g. Outcome types.

Add getFilterableHearingOutcomeStates which returns the number
of matches for each hearing defendant outcome state: new, in_progress and resulted
e.g. IN_PROGRESS: 2  means there are 2 hearing defendant outcomes that have
the in_progress state.

This will be used for the v2 hearing-defendant-outcomes endpoint

* Add tests for v2 endpoint

* Add v2 (version 2) endpoint for hearing outcomes

Add v2/courts/:code/hearing-defendant-outcomes endpoint

which provides a filtered and paginated response

* Update tests

* Set filter for outcome states to not allow multiple params

* Fix FilterHearingDefendantOutcomesServiceTest unit test