PIC-4138 Fix null pointer exception when trying to get offenders' defendants (#1278)
* Add null safety

Prevent null pointer exceptions by checking to see
if the offender.defendants is null or not. This is because we use
an optional offender.get() in order to hibernate.initialize / lazy
load defendant data

* Log when offender has no defendants associated

* Remove unrequired logging

* Refactor code

* Add unit test when offender has no defendants

Add unit test to prove that an empty optional is returned if
an offender has no defendants, which could be because the defendant
hasn't been saved to the database yet

* Add test if offender exists but no associated defendant

Add unit test to check that an offender can exist without
the defendant existing in the database. Also that it returns
the existing offender, if so.