Fix audit table issue in preprod
When a record is deleted in the original table, the audit table
does an insert with null values in the columns. However, there's a
not null constraint on the prep_status column which causes an exception.
This configuration should stop the need to add null values for a deltion.

This fixes the issue below:

org.postgresql.util.PSQLException: ERROR: null value in column "prep_status" of relation "hearing_defendant_aud" violates not-null constraint