PIC-4471 Update SAR services structure
Update the SAR services to return a json in the format:

cases -> list of HearingSarResponse -> HearingNotesSarResponse
cases -> list of HearingSarResponse -> HearingOutcomeSarResponse
case -> comments

This better explains the relationship between the data that we return
to the SAR team so that they are able to decipher it.

Don't return deleted, draft or legacy data.