Fix precision and scale
At the moment, prod and preprod do not have a precision or scale On
shorttermcustodypredictorscore in the offence table. However, the audit
table has a default 38 precision and 2 scale which is invalid.

Add migration to set the precision and scale of shorttermcustodypredictorscore
to 21,19.

This fixes invalid numeric types in preprod and prod audit tables too.

This fixes flaky tests too